requirements.txt
mlflow
opencv-python
Pillow
Django
tensorflow-cpu
openai
djangorestframework
djangorestframework-simplejwt
django-cors-headers